backend/src/requirements.txt
redis
Flask
Flask-Cors
flask-restx
PyJWT==1.7.1
Flask-SQLAlchemy
Flask-Migrate
passlib
cryptography
cffi
docopt
gunicorn==20.1.0
Flask-Script
python-dateutil
psycopg2-binary
xlrd
celery
eventlet==0.30.2
requests
sentry-sdk[flask]
webargs
python-keycloak
dataclasses